add screen folder
add a user dashboard file
